requirements.txt
asyncio
faiss_cpu
goal_plot
matplotlib
mplsoccer
networkx
numpy
pandas
plotly
psycopg2-binary
python-dotenv
requests
reportlab
seaborn
scikit-learn
scipy
soccerplots
sqlalchemy
statsbombpy
streamlit